Some cleanup work
