+++ cue.mod/pkg/github.com/TangoGroup/aws/uswest2/Elasticsearch.cue
@@ -85,7 +85,7 @@ import "github.com/TangoGroup/aws/fn"
 		DeletionPolicy?:      "Delete" | "Retain"
 		UpdateReplacePolicy?: "Delete" | "Retain"
 		UpdatePolicy?: [string]: _
-		Metadata?: [string]:     _
+		Metadata?: [string]: _
 		Condition?: string
 	}
 }
