+++ cue.mod/pkg/github.com/TangoGroup/aws/uswest2/Lambda.cue
@@ -27,7 +27,7 @@ import (
 		DeletionPolicy?:      "Delete" | "Retain"
 		UpdateReplacePolicy?: "Delete" | "Retain"
 		UpdatePolicy?: [string]: _
-		Metadata?: [string]:     _
+		Metadata?: [string]: _
 		Condition?: string
 	}
 	#CodeSigningConfig: {
